+++ sdk/browser/CHANGELOG.md
@@ -1,3 +1,5 @@
+# release 4.2.4 (2021-04-28)
+* custom messages support for passwordless login
 # release 4.2.3 (2020-04-01)
  * add optional pagination to `AffinityWallet.getCredentials`
  * add new method `AffinityWallet.getCredentialByIndex` that returns credentials given at the index